Struct solar_model::solar_surface::SolarSurface

pub struct SolarSurface {
    points: Vec<Point3D>,
    pub normal: Vector3D,
    nrays: usize,
}

Structure that can help calculate solar radiation

It contains the normal of the original Surface and the points randomly sampled in each surface.

Fields
points: Vec<Point3D>normal: Vector3Dnrays: usize
Implementations

impl SolarSurface

const DELTA: f64

Offset for the starting point of the rays.

pub fn new(nrays: usize, polygon: &Polygon3D) -> Self

fn calc_back_solar_dc_matrix(
list: &[SolarSurface],
scene: &Scene,
dc_factory: &DCFactory
) -> Matrix

Receives a list of SolarSurface objects as well as the Scene containing them and calculates the back Daylight Coefficient Matrix that can be used for estimating the incident solar radiation in W/m2. The options for this calculation are contained in the DCFactory used as input.

fn calc_front_solar_dc_matrix(
list: &[SolarSurface],
scene: &Scene,
dc_factory: &DCFactory
) -> Matrix

Receives a list of SolarSurface objects as well as the Scene containing them and calculates the front Daylight Coefficient Matrix that can be used for estimating the incident solar radiation in W/m2. The options for this calculation are contained in the DCFactory used as input.

pub fn get_front_solar_dc_matrix(
list: &[SolarSurface],
path: &Option<String>,
scene: &Scene,
dc_factory: &DCFactory
) -> Result<Matrix, String>

Gets the front Daylight Coefficient Matrix that can be used for estimating the incident solar radiation in W/m2. If the path exists, then it will attempt to read the matrix in that file. If it does not exist, it will calculate it and write it down there. If no path is given, it will just calculate and not save it anywhere.

pub fn get_back_solar_dc_matrix(
list: &[SolarSurface],
path: &Option<String>,
scene: &Scene,
dc_factory: &DCFactory
) -> Result<Matrix, String>

Gets the back Daylight Coefficient Matrix that can be used for estimating the incident solar radiation in W/m2. If the path exists, then it will attempt to read the matrix in that file. If it does not exist, it will calculate it and write it down there. If no path is given, it will just calculate and not save it anywhere.

pub fn make_fenestrations(
list: &[Rc<Fenestration>],
state: &mut SimulationStateHeader,
n_rays: usize
) -> Vec<SolarSurface>

Builds a set of SolarSurfaces from Fenestrations

Adds the necessary elements to the SimulationStateHeader

pub fn make_surfaces(
list: &[Rc<Surface>],
state: &mut SimulationStateHeader,
n_rays: usize
) -> Vec<SolarSurface>

Builds a set of SolarSurfaces from Surfaces

Adds the necessary elements to the `SimulationStateHeader

pub fn front_rays(&self) -> Vec<Ray3D>

Gets the front rays of a surface

pub fn back_rays(&self) -> Vec<Ray3D>

Gets the back rays of a surface

pub fn front_solar_irradiance(
&self,
scene: &Scene,
factory: &DCFactory
) -> Matrix

Calculates the Daylight Coefficient matrix for the front of a SolarSurface

pub fn back_solar_irradiance(
&self,
scene: &Scene,
factory: &DCFactory
) -> Matrix

Calculates the Daylight Coefficient matrix for the back of a SolarSurface